+++ app/website/models.py
@@ -1,0 +1,30 @@
+from . import db, ma
+
+
+# Models
+class User(db.Model):
+    user_id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(50), unique=True)
+    password = db.Column(db.String(32))
+    f_name = db.Column(db.String(32))
+    l_name = db.Column(db.String(32))
+    age = db.Column(db.Integer)
+
+    def __init__(self, username, password, first, last, age):
+        self.username = username
+        self.password = password
+        self.f_name = first
+        self.l_name = last
+        self.age = age
+
+
+# Schema
+class UsersSchema(ma.SQLAlchemyAutoSchema):
+    class Meta:
+        model = User
+        include_fk = True
+
+
+# Schemas init
+user_schema = UsersSchema()
+users_schema = UsersSchema(many=True)
